+++ R/matching_l1.R
@@ -5,6 +5,8 @@
 #' @param e_gps_std_pred a vector of predicted std of gps obtained by Machine learning methods.
 #' @param w the targeted single exposure levels.
 #' @param w_resid the standardized residuals for w.
+#' @param w_mx a vector with length 2, includes min(w), max(w).
+#' @param gps_mx a vector with length 2, includes min(gps), max(gps)
 #' @param scale a specified scale parameter to control the relative weight that is attributed to
 #' the distance measures of the exposure versus the GPS estimates (Default is 0.5).
 #' @param delta_n a specified caliper parameter on the exposure (Default is 1).
@@ -18,29 +20,35 @@ matching_l1 <- function(w,
                         e_gps_pred,
                         e_gps_std_pred,
                         w_resid,
+                        w_mx,
+                        gps_mx,
                         delta_n=1,
                         scale=0.5)
 {
-  w_new <- (w-e_gps_pred)/e_gps_std_pred
-  p.w <- approx(density(w_resid,na.rm = TRUE)$x,density(w_resid,na.rm = TRUE)$y,xout=w_new,rule=2)$y
+  w_new <- compute_resid(w, e_gps_pred, e_gps_std_pred)
+  p.w <- compute_density(w_resid, w_new)
 
-  w.min <- min(dataset[["w"]],na.rm=T)
-  w.max <- max(dataset[["w"]],na.rm=T)
-  gps.min <- min(dataset[["gps"]],na.rm=T)
-  gps.max <- max(dataset[["gps"]],na.rm=T)
+  w.min <- w_mx[1]
+  w.max <- w_mx[2]
+  gps.min <- gps_mx[1]
+  gps.max <- gps_mx[2]
+
   ##
   dataset <- transform(dataset,
                        std.w = (w - w.min) / (w.max - w.min),
                        std.gps = (gps - gps.min) / (gps.max - gps.min))
+
   std.w <- (w - w.min) / (w.max - w.min)
   std.p.w <- (p.w - gps.min) / (gps.max - gps.min)
-  ##
+
   dataset.subset <- dataset[abs(dataset[["w"]] - w) <= (delta_n/2), ]
-  ##
-  wm <- apply(abs(outer(dataset.subset[["std.gps"]], std.p.w, `-`)) * scale,
-              2,
-              function(x) which.min(abs(dataset.subset[["std.w"]] - std.w) * (1 - scale) + x)
-  )
+
+  wm <- compute_closest_wgps(dataset.subset[["std.gps"]],
+                             std.p.w,
+                             dataset.subset[["std.w"]],
+                             std.w,
+                             scale)
+
   dp <- dataset.subset[wm,]
   return(dp)
   gc()